+++ examples/react-app/src/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {
   PlmgButton,
+  PlmgCard,
   PlmgHeader,
   PlmgPageContainer,
   PlmgSidebar,
@@ -47,7 +48,7 @@ ReactDOM.render(
         }}
         type={'button'}
       >
-        Button
+        PlmgButton
       </PlmgButton>
       <br />
       <PlmgButton
@@ -62,6 +63,23 @@ ReactDOM.render(
       <PlmgSvgIcon icon={'home'} size={'6em'} />
       PlmgSvgIcon home
       <br />
+      <PlmgCard
+        headerText="Header Text"
+        topActionIcon={'home'}
+        topActionLabel={'Top Action'}
+        bottomActionText="Click here"
+        bottomButtonAction={() =>
+          console.log('bottomButtonClicked event received')
+        }
+        topActionClicked={() => console.log('topActionClicked event received')}
+      >
+        <div slot="slot-1">
+          <h1>PlmgCard slot-1</h1>
+        </div>
+        <div slot="slot-2">
+          <h1>PlmgCard slot-2</h1>
+        </div>
+      </PlmgCard>
     </div>
   </PlmgPageContainer>,
   document.getElementById('root')