+++ packages/component-library-react/src/components.ts
@@ -9,6 +9,7 @@ import { defineCustomElements } from '@ducky/plumage/loader';
 
 defineCustomElements();
 export const PlmgButton = /*@__PURE__*/createReactComponent<JSX.PlmgButton, HTMLPlmgButtonElement>('plmg-button');
+export const PlmgCard = /*@__PURE__*/createReactComponent<JSX.PlmgCard, HTMLPlmgCardElement>('plmg-card');
 export const PlmgHeader = /*@__PURE__*/createReactComponent<JSX.PlmgHeader, HTMLPlmgHeaderElement>('plmg-header');
 export const PlmgPageContainer = /*@__PURE__*/createReactComponent<JSX.PlmgPageContainer, HTMLPlmgPageContainerElement>('plmg-page-container');
 export const PlmgSidebar = /*@__PURE__*/createReactComponent<JSX.PlmgSidebar, HTMLPlmgSidebarElement>('plmg-sidebar');