+++ packages/component-library/src/components.d.ts
@@ -61,6 +61,24 @@ export namespace Components {
          */
         "type": PlmgButtonType;
     }
+    interface PlmgCard {
+        /**
+          * Define card's bottom button text.  If a text is provided, the button will be displayed. By default, when no text is provided, the button is hidden.
+         */
+        "bottomActionText": string | undefined;
+        /**
+          * Define card's header text.  If a headerText or an topActionIcon is provided, the heading will be displayed. By default, when no headerText nor topActionIcon is provided, the heading is hidden.
+         */
+        "headerText": string | undefined;
+        /**
+          * Define card's header icon, used as a top action for the card.  If a headerText or an topActionIcon is provided, the heading will be displayed with the icon button on the right. By default, when no headerText nor topActionIcon is provided, the heading is hidden.
+         */
+        "topActionIcon": string | undefined;
+        /**
+          * Define top action's label, used to enable assistive technology for the top action button.  You must provide a topActionLabel when providing a topActionIcon.
+         */
+        "topActionLabel": string | undefined;
+    }
     interface PlmgHeader {
         /**
           * Invoke this method to reveals the "expand" icon and update the margin left
@@ -147,6 +165,12 @@ declare global {
         prototype: HTMLPlmgButtonElement;
         new (): HTMLPlmgButtonElement;
     };
+    interface HTMLPlmgCardElement extends Components.PlmgCard, HTMLStencilElement {
+    }
+    var HTMLPlmgCardElement: {
+        prototype: HTMLPlmgCardElement;
+        new (): HTMLPlmgCardElement;
+    };
     interface HTMLPlmgHeaderElement extends Components.PlmgHeader, HTMLStencilElement {
     }
     var HTMLPlmgHeaderElement: {
@@ -179,6 +203,7 @@ declare global {
     };
     interface HTMLElementTagNameMap {
         "plmg-button": HTMLPlmgButtonElement;
+        "plmg-card": HTMLPlmgCardElement;
         "plmg-header": HTMLPlmgHeaderElement;
         "plmg-page-container": HTMLPlmgPageContainerElement;
         "plmg-sidebar": HTMLPlmgSidebarElement;
@@ -240,6 +265,32 @@ declare namespace LocalJSX {
           * Define button's type  Allowed values:   - button   - submit   - reset  Default: button
          */
         "type"?: PlmgButtonType;
+    }
+    interface PlmgCard {
+        /**
+          * Define card's bottom button text.  If a text is provided, the button will be displayed. By default, when no text is provided, the button is hidden.
+         */
+        "bottomActionText"?: string | undefined;
+        /**
+          * Define card's header text.  If a headerText or an topActionIcon is provided, the heading will be displayed. By default, when no headerText nor topActionIcon is provided, the heading is hidden.
+         */
+        "headerText"?: string | undefined;
+        /**
+          * The event "bottomActionClicked" is triggered when the bottom action button is clicked.
+         */
+        "onBottomActionClicked"?: (event: CustomEvent<MouseEvent>) => void;
+        /**
+          * The event "topActionClicked" is triggered when the top action button is clicked.
+         */
+        "onTopActionClicked"?: (event: CustomEvent<MouseEvent>) => void;
+        /**
+          * Define card's header icon, used as a top action for the card.  If a headerText or an topActionIcon is provided, the heading will be displayed with the icon button on the right. By default, when no headerText nor topActionIcon is provided, the heading is hidden.
+         */
+        "topActionIcon"?: string | undefined;
+        /**
+          * Define top action's label, used to enable assistive technology for the top action button.  You must provide a topActionLabel when providing a topActionIcon.
+         */
+        "topActionLabel"?: string | undefined;
     }
     interface PlmgHeader {
         /**
@@ -321,6 +372,7 @@ declare namespace LocalJSX {
     }
     interface IntrinsicElements {
         "plmg-button": PlmgButton;
+        "plmg-card": PlmgCard;
         "plmg-header": PlmgHeader;
         "plmg-page-container": PlmgPageContainer;
         "plmg-sidebar": PlmgSidebar;
@@ -333,6 +385,7 @@ declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
             "plmg-button": LocalJSX.PlmgButton & JSXBase.HTMLAttributes<HTMLPlmgButtonElement>;
+            "plmg-card": LocalJSX.PlmgCard & JSXBase.HTMLAttributes<HTMLPlmgCardElement>;
             "plmg-header": LocalJSX.PlmgHeader & JSXBase.HTMLAttributes<HTMLPlmgHeaderElement>;
             "plmg-page-container": LocalJSX.PlmgPageContainer & JSXBase.HTMLAttributes<HTMLPlmgPageContainerElement>;
             "plmg-sidebar": LocalJSX.PlmgSidebar & JSXBase.HTMLAttributes<HTMLPlmgSidebarElement>;